refactor: restructure backend code to server/
