+++ babel.config.js
@@ -1,0 +1,6 @@
+export default {
+  presets: [
+    ['@babel/preset-env', { useBuiltIns: 'usage' }],
+    '@babel/preset-react',
+  ],
+}
