+++ index.js
@@ -1,5 +1,5 @@
-import app from './app.js'
-import config from './utils/config.js'
+import app from './server/index.js'
+import config from './server/utils/config.js'
 
 app.listen(config.PORT, () => {
   console.log(`Server's running on port ${config.PORT}`)
